syncthing/requirements.txt
docker-py==1.10.3
docker-pycreds==0.2.1
pypiwin32==219
requests==2.10.0
six==1.10.0
websocket-client==0.37.0
wheel==0.24.0